+++ netdiff/__init__.py
@@ -12,3 +12,9 @@ def get_version():
         if VERSION[3] != 'final':
             version = '%s %s' % (version, VERSION[3])
     return version
+
+
+from .parsers.olsr import OlsrParser  # noqa
+from .parsers.batman import BatmanParser  # noqa
+from .parsers.netjson import NetJsonParser  # noqa
+from .utils import diff  # noqa