+++ CHANGES.rst
@@ -4,4 +4,8 @@ Changelog
 Version 0.1 [unreleased]
 ------------------------
 
-- TODO
+- Added ``OlsrParser``
+- Added ``BatmanParser``
+- Added ``NetJsonParser``
+- Added ``json`` method to ``BaseParser`` which returns valid NetJSON
+- Added ``diff`` function